requirements.txt
speechrecognition
pyttsx3
pygame
requests
google-generativeai
newsapi-python
pyaudio
